# contracts/openapi/pie-control-plane-v1.yaml
openapi: 3.1.2
info:
  title: Pie Control Plane API
  version: 0.1.0
  description: Initial executable contract for instance discovery, session, organizations, projects, and work items.
servers:
  - url: https://api.pielab.ai
security:
  - PieOidc: []
...
paths:
...
  /v1/organizations/{organizationId}/channels/{channelId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Collaboration]
      operationId: getChannel
      responses:
        '200':
          description: Channel representation.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/channel.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Collaboration]
      operationId: updateChannel
      description: Updates channel metadata or archives/restores a normal channel using optimistic concurrency.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../schemas/resources/channel-update.v1.schema.json
      responses:
        '200':
          description: Updated channel.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/channel.v1.schema.json
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/channels/{channelId}/members:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Collaboration]
      operationId: listChannelMembers
      responses:
        '200':
          description: Current channel roster.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: ../schemas/resources/channel-member.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Collaboration]
      operationId: addChannelMember
      description: >-
        Adds an org member to a normal channel (channel.manage). Rejected on a DM
        (409) — a DM's roster is fixed at its participants. Idempotent (204).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              additionalProperties: false
              properties:
                userId:
                  $ref: ../schemas/common/opaque-id.v1.schema.json
      responses:
        '204':
          description: Member added (or already present).
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/channels/{channelId}/audit:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Collaboration]
      operationId: listChannelAudit
      description: Lists the latest 200 channel and message governance events for channel managers.
      responses:
        '200':
          description: Recent governance activity, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: ../schemas/resources/channel-audit-entry.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/channels/{channelId}/export:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    get:
      tags: [Collaboration]
      operationId: exportChannelMessages
      description: Exports up to 10,000 channel messages for channel managers.
      responses:
        '200':
          description: A point-in-time JSON export including tombstone audit metadata.
          content:
            application/json:
              schema:
                $ref: ../schemas/resources/channel-message-export.v1.schema.json
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/channels/{channelId}/retention:apply:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    post:
      tags: [Collaboration]
      operationId: applyChannelRetention
      description: Redacts expired message and revision bodies using the channel retention policy.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Retention applied successfully.
          content:
            application/json:
              schema:
                type: object
                required: [ok, redactedCount]
                additionalProperties: false
                properties:
                  ok:
                    const: true
                  redactedCount:
                    type: integer
                    minimum: 0
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
  /v1/organizations/{organizationId}/channels/{channelId}/members/{userId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - name: channelId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
      - name: userId
        in: path
        required: true
        schema:
          $ref: ../schemas/common/opaque-id.v1.schema.json
    delete:
      tags: [Collaboration]
      operationId: removeChannelMember
      description: Removes a member from a normal channel while preserving at least one owner.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Member removed (or was already absent).
        '409':
          $ref: '#/components/responses/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    PieOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.pielab.ai/realms/pie/.well-known/openid-configuration
      description: The actual issuer is discovered from /.well-known/pie for each instance.
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        $ref: ../schemas/common/opaque-id.v1.schema.json
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
  headers:
    ETag:
      description: Strong entity tag for the returned resource version.
      required: true
      schema:
        type: string
        pattern: '^"[^" ]+"$'
...
  responses:
    Problem:
      description: RFC 9457 problem with a stable Pie error code.
      content:
        application/problem+json:
          schema:
            $ref: ../schemas/common/problem-details.v1.schema.json